requirements.txt
APScheduler==3.10.4
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.40
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5
MarkupSafe==2.1.3
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
smmap==5.0.1
tzdata==2023.3
tzlocal==5.1
Werkzeug==3.0.0
